dags: Change file naming conventions.
Use the metadata file name as the prefix to the output files instead
of the name of the original input file.
Use created_at column in database in order to create a subfolder
under `output/file_id`, using created_at from the database allows
us to have the same created_at date for both the sessions.