dags: Use channel index in the channel list as speaker_id.
The user has to specify the speaker_id in the hybrids section instead
of channel. If `channel` field is present instead of `speaker_id` i.e.
in old json, we default to the value of channel but use it as the speaker id
to support old files.
NOTE: REMOVE THE DEFAULTING TO CHANNEL ONCE WE NO LONGER NEED TO SUPPORT THOSE
OLD FILES.